API Reference

API Reference for LlamaCpp.

LlamaCpp.download_model

LlamaCpp.embeddings

LlamaCpp.llama_eval

LlamaCpp.logits

LlamaCpp.run_chat

LlamaCpp.run_llama

LlamaCpp.run_server

LlamaCpp.token_to_str

LlamaCpp.tokenize

LlamaCpp.download_model — Method

download_model(url::AbstractString; dir::AbstractString="models")

Downloads a model specified by url from the HuggingFace Hub into dir directory and returns the model_path to the downloaded file. If the dir directory does not exist, it will be created.

Note: Currently allows only models in the GGUF format (expects the URL to end with .gguf).

See HuggingFace Model Hub for a list of available models.

Examples

# Download the Rocket model (~1GB)
url = "https://huggingface.co/ikawrakow/various-2bit-sota-gguf/resolve/main/rocket-3b-2.76bpw.gguf"
model = download_model(url) 
# Output: "models/rocket-3b-2.76bpw.gguf"

LlamaCpp.embeddings — Method

embeddings(ctx) -> Vector{Float32}

Return the embedding, a vector of length ctx.n_embd.

LlamaCpp.llama_eval — Method

llama_eval(ctx::LlamaContext, tokens; n_past, n_threads=1)

LlamaCpp.logits — Method

logits(ctx::LlamaContext) -> Vector{Float32}

Return the logits (unnormalised probabilities) for each token, a vector of length ctx.n_vocab.

LlamaCpp.run_chat — Method

run_chat(; model::AbstractString, prompt::AbstractString="", nthreads::Int=Threads.nthreads(), n_gpu_layers::Int=99, ctx_size::Int=2048, args=``)

Opens an interactive console for the model and runs in "instruction" mode (especially useful for Alpaca-based models). prompt, as the first message, is often used to provide instruction about the upcoming interactions (eg, style, tone, roles).

Wait for model to reply and then type your response. Press Enter to send the message to the model.

Interrupt the chat with Ctrl+C

See the full documentation for more details.

Arguments

model: path to the model to be used

prompt: prompt to be used. Most models expected these to be formatted in a specific way. Defaults to an empty string

nthreads: number of threads to use. Defaults to the number of available threads

n_gpu_layers: number of layers to offload on the GPU (a.k.a. ngl in llama.cpp). Requires more VRAM on your GPU but can speed up inference. Set to 0 to run inference on CPU-only. Defaults to 99 (=practically all layers)

ctx_size: context size, ie, how big can the prompt/inference be. Defaults to 2048 (but most models allow 4,000 and more)

Note: If you get odd responses AND you're using an instruction-tuned ("fine-tuned"), it might be that the format of your prompt is not correct. See HuggingFace's model documentation for the correct prompt format or use a library that will do this for you (eg, PromptingTools.jl)

See also: run_llama, run_server

LlamaCpp.run_llama — Method

run_llama(; model::AbstractString, prompt::AbstractString="", nthreads::Int=1, n_gpu_layers::Int=99, ctx_size::Int=2048, args=``)

Runs prompt through the model provided and returns the result. This is a single-turn version of run_chat.

See the full documentation for more details.

Arguments

model: path to the model to be used

prompt: prompt to be used. Most models expected these to be formatted in a specific way. Defaults to an empty string

nthreads: number of threads to use. Defaults to the number of available threads

n_gpu_layers: number of layers to offload on the GPU (a.k.a. ngl in llama.cpp). Requires more VRAM on your GPU but can speed up inference. Set to 0 to run inference on CPU-only. Defaults to 99 (=practically all layers)

ctx_size: context size, ie, how big can the prompt/inference be. Defaults to 2048 (but most models allow 4,000 and more)

Note: If you get odd responses AND you're using an instruction-tuned ("fine-tuned"), it might be that the format of your prompt is not correct. See HuggingFace's model documentation for the correct prompt format or use a library that will do this for you (eg, PromptingTools.jl)

See also: run_chat, run_server

LlamaCpp.run_server — Method

run_server(; model::AbstractString, host::AbstractString="127.0.0.1", port::Int=10897, nthreads::Int=Threads.nthreads(), 
n_gpu_layers::Int=99, ctx_size::Int=2048, args=``)

Starts a simple HTTP server with the model provided.

Open http://{host}:{port} in your browser to interact with the model or use an HTTP client to send requests to the server.

Interrupt the server with Ctrl+C.

Arguments

model: path to the model to be used

host: host address to bind to. Defaults to "127.0.0.1"

port: port to listen on. Defaults to 10897

nthreads: number of threads to use. Defaults to the number of available threads

n_gpu_layers: number of layers to offload on the GPU (a.k.a. ngl in llama.cpp). Requires more VRAM on your GPU but can speed up inference. Set to 0 to run inference on CPU-only. Defaults to 99 (=practically all layers)

ctx_size: context size, ie, how big can the prompt/inference be. Defaults to 2048 (but most models allow 4,000 and more)

embeddings: whether to allow generating of embeddings. Defaults to false. Note: Embeddings are not supported by all models and it might break the server!

args: additional arguments to pass to the server

See the full documentation for more details.

Example

```julia using LlamaCpp

Download a model from HuggingFace, eg, Phi-2.

See details here

using Downloads model = joinpath("models", "dolphin-26-phi-2.Q6K.gguf") mkpath(dirname(model)) # ensure the folder exists Downloads.download("https://huggingface.co/TheBloke/dolphin-26-phi-2-GGUF/resolve/main/dolphin-26-phi-2.Q6_K.gguf", model)

go make a cup of tea while you wait... this is a 2.3GB download

Start the server

run_server(; model)

LlamaCpp.token_to_str — Method

token_to_str(ctx, token_id) -> String

String representation for token token_id.

LlamaCpp.tokenize — Method

tokenize(ctx :: LlamaContext, text :: AbstractString) -> Vector{llama_token}

Tokenizes text according to the LlamaContext ctx and returns a Vector{llama_token}, with llama_token == Int32.